+++ package.json
@@ -73,8 +73,7 @@
     "cp:flow": "cp node_modules/@nodeproto/configproto/.flowconfig .",
     "eslint:fix": "pnpm eslint -- --fix",
     "eslint": "eslint 'src/**/*'",
-    "nvm:latest": "nvm install node --reinstall-packages-from=current",
-    "post:nvm:latest": "nvm alias default node",
+    "pkg:deps": "pnpm exec ultra --info",
     "test:help": "uvu --help",
     "test:ing": "clear; pnpm test; chokidar \"src/**/*(?:.test)?.(m|c)?js\" -c \"clear; pnpm test\"",
     "test": "NODE_OPTIONS=\"$npm_package_config_NODE_OPTIONS\" uvu ./src test.mjs",
