+++ packages/libraries/configproto/package.json
@@ -75,7 +75,9 @@
     "post:nvm:latest": "nvm alias default node",
     "eslint": "eslint 'src/**/*'",
     "eslint:fix": "pnpm eslint -- --fix",
-    "cp:flow": "cp node_modules/@nodeproto/configproto/.flowconfig ."
+    "cp:flow": "cp node_modules/@nodeproto/configproto/.flowconfig .",
+    "config:list": "pnpm config list --json",
+    "pkg:deps": "pnpm exec ultra --info"
   },
   "jsync": {
     "root": false
@@ -197,7 +199,9 @@
     "webpack": "5.51.1",
     "webpack-dev-middleware-2": "5.0.2",
     "webpack-hot-middleware-2": "2.25.2",
-    "yaml-eslint-parser": "0.4.1"
+    "yaml-eslint-parser": "0.4.1",
+    "@changesets/cli": "2.17.0",
+    "ultra-runner": "3.10.5"
   },
   "peerDependencies": {
     "@babel/cli": "7.14.8",
@@ -378,11 +382,18 @@
   "repository": "git://github.com/noahehall/nodeproto.git",
   "type": "module",
   "dependencies": {
-    "@nodeproto/envproto": "workspace:0.0.0",
     "chai-as-promised": "7.1.1",
     "chai-exclude": "2.0.3",
     "chai-nock": "1.3.0",
     "nock": "13.1.3",
     "sinon": "11.1.1"
-  }
+  },
+  "ultra": {
+    "concurrent": [
+      "eslint",
+      "test",
+      "cp:flow"
+    ]
+  },
+  "packageManager": "pnpm@6.15.1"
 }
