+++ packages/libraries/envproto/package.json
@@ -32,7 +32,9 @@
   "devDependencies": {
     "@nodeproto/jsync": "workspace:*",
     "chokidar-cli": "3.0.0",
-    "uvu": "0.5.1"
+    "uvu": "0.5.1",
+    "@changesets/cli": "2.17.0",
+    "ultra-runner": "3.10.5"
   },
   "scripts": {
     "about": "npm run-script && echo 'available via `bin/FILENAME`:\n'; ls bin/*.sh || exit 0",
@@ -44,7 +46,9 @@
     "eslint:fix": "pnpm eslint -- --fix",
     "test": "NODE_OPTIONS=\"$npm_package_config_NODE_OPTIONS\" uvu ./src test.mjs",
     "test:help": "uvu --help",
-    "test:ing": "pnpm test; chokidar \"src/**/*(?:.test)?.(m|c)?js\" -c \"clear; pnpm test\""
+    "test:ing": "clear; pnpm test; chokidar \"src/**/*(?:.test)?.(m|c)?js\" -c \"clear; pnpm test\"",
+    "config:list": "pnpm config list --json",
+    "pkg:deps": "pnpm exec ultra --info"
   },
   "peerDependencies": {
     "@nodeproto/jsync": "workspace:*"
@@ -89,5 +93,13 @@
     "darwin",
     "linux"
   ],
-  "repository": "git://github.com/noahehall/nodeproto.git"
+  "repository": "git://github.com/noahehall/nodeproto.git",
+  "ultra": {
+    "concurrent": [
+      "eslint",
+      "test",
+      "cp:flow"
+    ]
+  },
+  "packageManager": "pnpm@6.15.1"
 }
