+++ packages/shared/index.mjs
@@ -1,0 +1,1 @@
+export * as wtf from './src/wtf/index.mjs';
