+++ packages/shared/src/wtf/index.mjs
@@ -1,0 +1,2 @@
+export * from './readdir.mjs';
+export * from './dirnamefilename.mjs';
