+++ packages/tools/inception/package.json
@@ -21,7 +21,9 @@
     "eslint:fix": "pnpm eslint -- --fix",
     "cp:flow": "cp node_modules/@nodeproto/configproto/.flowconfig .",
     "test:help": "uvu --help",
-    "test:ing": "clear; pnpm test; chokidar \"src/**/*(?:.test)?.(m|c)?js\" -c \"clear; pnpm test\""
+    "test:ing": "clear; pnpm test; chokidar \"src/**/*(?:.test)?.(m|c)?js\" -c \"clear; pnpm test\"",
+    "config:list": "pnpm config list --json",
+    "pkg:deps": "pnpm exec ultra --info"
   },
   "dependencies": {
     "@nodeproto/wtf": "workspace:*",
@@ -85,10 +87,18 @@
   "devDependencies": {
     "@nodeproto/jsync": "workspace:*",
     "ultra-runner": "3.10.5",
-    "uvu": "0.5.1"
+    "uvu": "0.5.1",
+    "@changesets/cli": "2.17.0"
   },
   "peerDependencies": {
     "@nodeproto/jsync": "workspace:*"
   },
-  "packageManager": "pnpm@6.15.1"
+  "packageManager": "pnpm@6.15.1",
+  "ultra": {
+    "concurrent": [
+      "eslint",
+      "test",
+      "cp:flow"
+    ]
+  }
 }
