+++ packages/tools/jsync/package.json
@@ -17,13 +17,16 @@
     "jsync": "node ./bin/jsync.js",
     "test": "NODE_OPTIONS=\"$npm_package_config_NODE_OPTIONS\" uvu ./src test.mjs",
     "test:help": "uvu --help",
-    "test:ing": "clear; pnpm test; chokidar \"src/**/*(?:.test)?.(m|c)?js\" -c \"clear; pnpm test\""
+    "test:ing": "clear; pnpm test; chokidar \"src/**/*(?:.test)?.(m|c)?js\" -c \"clear; pnpm test\"",
+    "config:list": "pnpm config list --json",
+    "pkg:deps": "pnpm exec ultra --info"
   },
   "dependencies": {
-    "@nodeproto/wtf": "workspace:*",
+    "@nodeproto/shared": "workspace:0.0.0",
     "env-cmd": "10.1.0",
     "eslint": "7.32.0",
     "flow-bin": "0.157.0",
+    "fs-extra": "10.0.0",
     "ultra-runner": "3.10.5"
   },
   "jsync": {
@@ -81,8 +84,16 @@
   },
   "repository": "git://github.com/noahehall/nodeproto.git",
   "devDependencies": {
+    "@changesets/cli": "2.17.0",
     "ultra-runner": "3.10.5",
     "uvu": "0.5.1"
   },
-  "packageManager": "pnpm@6.15.1"
+  "packageManager": "pnpm@6.15.1",
+  "ultra": {
+    "concurrent": [
+      "eslint",
+      "test",
+      "cp:flow"
+    ]
+  }
 }
