+++ packages/tools/wtf/package.json
@@ -23,8 +23,6 @@
     "jsonc-simple-parser": "2.2.0",
     "memfs": "3.2.2",
     "mkdirp": "1.0.4",
-    "picomatch": "2.3.0",
-    "shelljs": "0.8.4",
     "symlink-dir": "5.0.1",
     "ultra-runner": "3.10.5",
     "yargs": "17.0.1"
@@ -73,15 +71,15 @@
   ],
   "scripts": {
     "about": "npm run-script && echo 'available via `bin/FILENAME`:\n'; ls bin/*.sh || exit 0",
-    "upgrade": "pnpm upgrade -L",
-    "nvm:latest": "nvm install node --reinstall-packages-from=current",
-    "post:nvm:latest": "nvm alias default node",
-    "eslint": "eslint 'src/**/*'",
+    "cp:flow": "cp node_modules/@nodeproto/configproto/.flowconfig .",
     "eslint:fix": "pnpm eslint -- --fix",
-    "test": "NODE_OPTIONS=\"$npm_package_config_NODE_OPTIONS\" uvu ./src test.mjs",
+    "eslint": "eslint 'src/**/*'",
     "test:help": "uvu --help",
     "test:ing": "clear; pnpm test; chokidar \"src/**/*(?:.test)?.(m|c)?js\" -c \"clear; pnpm test\"",
-    "cp:flow": "cp node_modules/@nodeproto/configproto/.flowconfig ."
+    "test": "NODE_OPTIONS=\"$npm_package_config_NODE_OPTIONS\" uvu ./src test.mjs",
+    "upgrade": "pnpm upgrade -L",
+    "config:list": "pnpm config list --json",
+    "pkg:deps": "pnpm exec ultra --info"
   },
   "config": {
     "NODE_OPTIONS": "--pending-deprecation --experimental-wasm-modules --trace-warnings --experimental-top-level-await --enable-source-maps --experimental-vm-modules --experimental-worker --experimental-import-meta-resolve --experimental-report --trace-exit --heapsnapshot-near-heap-limit=3 --trace-uncaught --deprecation --trace-deprecation --no-insecure-http-parser --report-uncaught-exception --use-largepages=on --warnings --experimental-abortcontroller --report-on-fatalerror --experimental-json-modules --trace-sigint --abort-on-uncaught-exception --es-module-specifier-resolution=node --experimental-specifier-resolution=node --experimental-modules",
@@ -91,11 +89,18 @@
     "BABEL_SHOW_CONFIG_FOR": "e.g. BABEL_SHOW_CONFIG_FOR=absolute/path/to/file somePkgJsonScriptName"
   },
   "devDependencies": {
-    "@nodeproto/configproto": "workspace:0.0.0",
+    "@changesets/cli": "2.17.0",
     "@nodeproto/jsync": "workspace:0.0.0",
     "eslint": "7.32.0",
-    "uvu": "0.5.1",
-    "ultra-runner": "3.10.5"
+    "ultra-runner": "3.10.5",
+    "uvu": "0.5.1"
   },
-  "packageManager": "pnpm@6.15.1"
+  "packageManager": "pnpm@6.15.1",
+  "ultra": {
+    "concurrent": [
+      "eslint",
+      "test",
+      "cp:flow"
+    ]
+  }
 }
