+++ packages/tools/wtf/src/dirs.mjs
@@ -1,24 +1,13 @@
-import { fileURLToPath } from 'url';
+
 import { homedir, tmpdir } from 'os';
+import { wtf as wtfShared } from '@nodeproto/shared';
 
 import cPath from 'contains-path';
 import Dirent from 'dirent';
 import globalDirs from 'global-dirs';
-import JSONC from 'jsonc-simple-parser';
 import path from 'path';
-import picomatch from 'picomatch';
-import Readdir from '@folder/readdir';
-import shelljs from 'shelljs';
 import symlinkDir from 'symlink-dir';
 import xdg from '@folder/xdg';
-
-const isCjs = typeof require === 'undefined';
-const filename = (importMetaUrl = import.meta?.url) => isCjs
-  ? fileURLToPath(importMetaUrl)
-  : __filename;
-const dirname = (importMetaUrl = import.meta?.url) => isCjs
-  ? path.dirname(filename(importMetaUrl))
-  : __dirname;
 
 /**
  * gets cross-platform directories
@@ -32,43 +21,8 @@ const dirname = (importMetaUrl = import.meta?.url) => isCjs
  * @TODO https://www.npmjs.com/package/unixify
  * @return {object} properties and methods supporting filesystem management
  */
+
 export const getDirs = (overrides = {}) => {
-  const readdirOptions = {
-    absolute: true,
-    depth: 2,
-    dot: false,
-    // basename: e.g. $HOME, // required if absolute:false,
-    // filter: mustContainThisPath e.g. 'poop',
-    follow: true, // symlinks
-    // isMatch: someRegex, see filter
-  };
-
-  const readdir = async ({ dirpath, glob, ...opts }) => {
-    return Readdir(
-      dirpath,
-      { ...readdirOptions, ...opts, isMatch: glob ? file => picomatch(glob)(file.relative) : undefined }
-    );
-  };
-
-  const getFilePathAbs = async (dirpath = '.', glob) => (await readdir({ dirpath, glob }));
-
-  const getPkgJsonAbs = async (dirpath = '.', glob) => (await getFilePathAbs(dirpath, glob))[0];
-
-  const getPkgJson = async (dirpath = '.', glob = 'package\.json', interpreter = JSONC.parse) => {
-    const pkgJsonAbs = await getPkgJsonAbs(dirpath, glob);
-
-    return pkgJsonAbs
-      && {
-        file: interpreter(shelljs.cat(pkgJsonAbs)),
-        path: pkgJsonAbs,
-      }
-      || {};
-  };
-
-  // TODO
-  // review microsoft/jsonc-parser thing
-  const getPkgJsonc = async (dirpath = '.') => getPkgJson(dirpath, 'package\.jsonc');
-
   return {
     ...xdg({
       env: process.env,
@@ -78,19 +32,12 @@ export const getDirs = (overrides = {}) => {
       tempdir: tmpdir(),
     }),
     cPath,
-    dirname,
-    filename,
-    getFilePathAbs,
-    getPkgJson,
-    getPkgJsonAbs,
-    getPkgJsonc,
     globalDirs,
     inceptionStore: `${homedir()}/.node_modules`,
-    JSONC,
-    picomatch,
-    readdir,
-    shelljs,
     symlinkDir,
+
+    ...external,
+    ...internal,
 
     ...overrides,
   };
